feat(auth): Implement autentication cookie
GetUsers: This function receives a connect.Request of type pbv1.GetUsersRequest and returns a connect.Response of type pbv1.GetUsersResponse along with a possible error. In the function body, the a.app.Queries.GetUsers.Handle field is used to perform some logic for fetching users, and a response is constructed using connect.NewResponse. Then, the response and possible error are returned.

Logout: Remove cookie, this function receives a connect.Request of type pbv1.LogoutRequest and returns a connect.Response of type pbv1.LogoutResponse along with a possible error. In the function body, a response is created using connect.NewResponse with a success message, and some cookie-related logic is performed using the auth.Cookie type. Then, the response and possible error are returned.

Login: This function receives a connect.Request of type pbv1.LoginRequest and returns a connect.Response of type pbv1.LoginResponse along with a possible error. In the function body, the a.app.Commands.Login.Handle field is used to handle the login logic. Then, a response is created using connect.NewResponse with a success message, and some cookie-related logic is performed using the auth.Cookie type. Finally, the response and possible error are returned.

Implements: https://github.com/migerick